refactor tx eip4844
